# api/openapi/order-service.yaml
openapi: "3.0.3"
info:
  title: Order Service API
  version: "1.0.0"
  description: |
    REST API for placing and managing orders in the Vyndra platform.

    Orders flow through the following lifecycle:
    PENDING → CONFIRMED → SHIPPED → DELIVERED
    Any state except SHIPPED, DELIVERED, and CANCELLED can transition to CANCELLED.
  contact:
    name: Vyndra Engineering
    email: engineering@vyndra.io

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Orders
    description: Order management operations

paths:
  /orders:
    post:
      tags: [Orders]
      operationId: placeOrder
      summary: Place a new order
      description: Creates a new order in PENDING state and emits an order-created event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceOrderRequest'
      responses:
        "201":
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Orders]
      operationId: getOrdersByCustomer
      summary: Get orders for a customer
      description: Returns all orders belonging to the specified customer.
      parameters:
        - name: customerId
          in: query
          required: true
          description: The customer identifier
          schema:
            type: string
            example: cust-001
      responses:
        "200":
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'

  /orders/{id}:
    get:
      tags: [Orders]
      operationId: getOrder
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          description: The order UUID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Orders]
      operationId: cancelOrder
      summary: Cancel an order
      description: |
        Cancels the order if it has not yet been shipped.
        Emits an order-cancelled event on success.
      parameters:
        - name: id
          in: path
          required: true
          description: The order UUID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "400":
          description: Order cannot be cancelled in its current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /orders/{id}/status:
    patch:
      tags: [Orders]
      operationId: updateOrderStatus
      summary: Update the status of an order
      description: Advances the order through its lifecycle. Invalid transitions return 400.
      parameters:
        - name: id
          in: path
          required: true
          description: The order UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
      responses:
        "200":
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        "400":
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    PlaceOrderRequest:
      type: object
      required: [customerId, items]
      properties:
        customerId:
          type: string
          description: Identifier of the customer placing the order
          example: cust-001
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemRequest'

    OrderItemRequest:
      type: object
      required: [productId, productName, quantity, unitPrice]
      properties:
        productId:
          type: string
          description: Product identifier
          example: prod-42
        productName:
          type: string
          description: Human-readable product name
          example: Wireless Keyboard
        quantity:
          type: integer
          minimum: 1
          description: Quantity ordered
          example: 2
        unitPrice:
          type: number
          format: double
          minimum: 0.01
          description: Unit price in USD
          example: 49.99

    UpdateOrderStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'

    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique order identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        customerId:
          type: string
          description: Identifier of the customer who placed the order
          example: cust-001
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        totalAmount:
          type: number
          format: double
          description: Total order value
          example: 149.97
        currency:
          type: string
          description: ISO 4217 currency code
          example: USD
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the order was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the order was last updated

    OrderItemResponse:
      type: object
      properties:
        productId:
          type: string
          description: Product identifier
          example: prod-42
        productName:
          type: string
          description: Human-readable product name
          example: Wireless Keyboard
        quantity:
          type: integer
          description: Quantity ordered
          example: 2
        unitPrice:
          type: number
          format: double
          description: Unit price in USD
          example: 49.99
        totalPrice:
          type: number
          format: double
          description: Total price for this line item (quantity × unitPrice)
          example: 99.98

    OrderStatus:
      type: string
      enum: [PENDING, CONFIRMED, SHIPPED, DELIVERED, CANCELLED]
      description: Lifecycle status of an order

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Human-readable error description
        timestamp:
          type: string
          format: date-time